requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
httpx==0.27.2
python-multipart==0.0.12
python-dotenv==1.0.1
openai==1.51.0
